UI Integration, Part 1
Hopefully beautified by Thomas N. or not, idk yet. Will be talking to this mythical figure also known as "Jorrik Poll" to figure these out. If the game can tell in which scene its in, by number, than we can have a fully automated UI window with description and all. It would take maybe an hour of work to integrate everything. However, I also am going to die in 38 minutes. Buh-bye!